+++ changelog.md
@@ -21,6 +21,7 @@
 - Add API to disclosure API capabilities [#16](https://github.com/AVENTER-UG/mesos-m3s/pull/16) (thanks to [@itsoksarvesh](https://github.com/itsoksarvesh)).
 - Update K3s to Version 1.25.2
 - Move environment variable KUBERNETES_VERSION from bootstrap file into m3s framework.
+- Add Timezone support via env variable `TZ` [#17](https://github.com/AVENTER-UG/mesos-m3s/pull/17) (thanks to [@itsoksarvesh](https://github.com/itsoksarvesh)).
 
 ## v0.4.0